requirements.txt
google-genai>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
streamlit>=1.30.0,<2.0.0